+++ src/config.ts
@@ -1,6 +1,7 @@
 import * as fs from 'fs'
 import * as path from 'path'
 import * as os from 'os'
+import { isCallsignSSID, isCallsign, isSSID } from './utils'
 import { Keystore } from './Keystore' 
 
 export interface Config {
@@ -33,6 +34,7 @@ export const defaultConfig: Config = {
  * @returns void
  */
 export function save(config: Config, configPath?: string): void {
+    validate(config)
     const path = typeof configPath === 'string' ? configPath : defaultConfigPath
     fs.writeFileSync(path, JSON.stringify(config, null, 4))
 }
@@ -44,7 +46,9 @@ export function save(config: Config, configPath?: string): void {
  */
 export function load(configPath?: string): Config {
     const path = typeof configPath === 'string' ? configPath : defaultConfigPath
-    return JSON.parse(fs.readFileSync(path).toString('utf8'))
+    const conf: Config = JSON.parse(fs.readFileSync(path).toString('utf8'))
+    validate(conf)
+    return conf
 }
 
 /** Check if the config file (or any file) exists
@@ -55,6 +59,37 @@ export function load(configPath?: string): Config {
 export function exists(configPath?: string): boolean {
     const path = typeof configPath === 'string' ? configPath : defaultConfigPath
     return fs.existsSync(path)
+}
+
+/**
+ * @function validate
+ * @param config 
+ * @throws TypeError
+ */
+export function validate(config: Config): void {
+    
+    if (typeof config !== 'object') {
+        throw TypeError('config is not an object')
+    } else if (typeof config.version !== 'number') {
+        throw TypeError('version must be a number type')
+    } else if (typeof config.callsign !== 'string') {
+        throw TypeError('callsign must be a string type')
+    } else if (isCallsignSSID(config.callsign)) {
+        throw TypeError('callsign must be a valid callsign excluding an SSID')
+    } else if (!isCallsign(config.callsign)) {
+        throw TypeError('callsign must be a valid callsign')
+    } else if (!isSSID(config.ssid)) {
+        throw TypeError('ssid must be a number between 0 and 15')
+    } else if (typeof config.kissPort !== 'string') {
+        throw TypeError('kissPort must be a string type')
+    } else if (typeof config.kissBaud !== 'number') {
+        throw TypeError('kissBaud must be a number type')
+    } else if (typeof config.keystoreFile !== 'string') {
+        throw TypeError('keystoreFile must be a string type')
+    } else if (typeof config.signingKey !== 'undefined' 
+          && typeof config.signingKey !== 'string') {
+        throw TypeError('signingKey must be a string if it is defined')
+    }
 }
 
 /**